requirements.txt
Django~=4.1.1
djangorestframework
django-cors-headers
djangorestframework_simplejwt
